refactor: changed implementation
