+++ lib/pages/personasAceptadas.dart
